Better autocache for property
Only define one property (which should play nicely with documentation
and type checking) and wraps this around to do the auto cache.
A small single call (@autocache), with no need to keep track of
auto_cache properties and corresponding methods to generate them.
Also a single cache to deal with now.