Refactor with ConfigKeys and ConfigValues
No sense in ConfigKV holding both key and value when all the configs are
held in dictionaries with that very key.
Since I want to keep some functionality with keys I created a thin
wrapper around a key str.

Back to normal dict, which I think is easier and more easily modified