Move get_trait_typehint and account allow_none
If allow_none is true, we kinda need to typeset trait as if
it is Union[trait, None]
We could recurse in some way but we would need to change the value
allow_none to avoid infinite recursion it's messy.
I added a sort of hook function that runs before returning the result.
I previously already hesitated on creating one for various purposes so I
guess it can be useful.