Use context manager for errors
It allows to deal automatically with deciding wether to raise or not.
Not quite as flexible than try/except, but very easy to use.
Also easy to globally configure error logging (level, message, etc.).

It required to move resolve_config out of Scheme into ApplicationBase,
and make it a bound method, but this makes sense. We really only resolve
full configs from the application, and the application is always
instanciated.

It also necessitated to remove ApplicationBase.apply_cli_config and do
that earlier in the ConfigLoader. Because the user could change the way
the error catching business with a CLI argument or in a config file, and
this impacts operations happening in the ConfigLoader (most importantly
the resolving, and parsing for CLILoader).

So a quite crude section of code finds Application traits in the
retrieved config, without any fancy resolving. It's fine since we only
want the simplest keys (not nested, or ApplicationClass.traitname) so
no difficulties. Might need to account for aliases in the future.