Small overhaul of writers
Remove params argument. It puts pressure on the writer to do parameter
management (and we have no idea what kind of parameters are there).
Now that parameters can easily be update it is no hassle to do something
like:
``` python
dm.update_params(Y=2000)
dm.write(ds.sel(Y=2000))
```

Re-ordered the call tuple to have `[source, data]`, to be consistent
with other generics, and instead of using a typevar Call used generics
directly for WriterPluginAbstract. It can lead to problems in some
functions, like the target argument of write(). I kept it as simple as
possible and ignored mypy for override. Users will see the prescribed
signature, even if it clashes with the abstract one.