Go back to discovering module types
add that the type of the actual module used will be assigned to
DataManagerBase.<Module._TYPE_ATTR> so that users can access it
easily and subclass it. mypy is not happy about it. kinda dropping that
atm. maybe a plugin could help but i'm not even sure i can write it...

this is starting to look good user side, the dataset definitions are
still nice by having class def nested. And the thing is easier to wield
than the mixins. Some interplay between modules is still not perfect. I
can deal by having a type check at module init (this is the case for the
split writer). but maybe it could be done better ?