Change instanciation methods around.
Not really tested for now

Change the instanciation method to a class method. Seems relevant.
Code to configure the present instance is added directly to Application,
where we have a parent instance but not its children. Otherwise, this is
not a pattern that seems useful to keep around (configure self and
instanciate subschemes).

Put Application instanciation in hidden method for power users that may
want to delay instanciation.